refactor: 의존성 주입을 위한 AppContainer 추가
