+++ app/src/main/java/woowacourse/shopping/di/AppContainer.kt
@@ -1,0 +1,11 @@
+package woowacourse.shopping.di
+
+import woowacourse.shopping.data.CartRepositoryImpl
+import woowacourse.shopping.data.ProductRepositoryImpl
+import woowacourse.shopping.domain.CartRepository
+import woowacourse.shopping.domain.ProductRepository
+
+object AppContainer {
+    val cartRepository: CartRepository by lazy { CartRepositoryImpl() }
+    val productRepository: ProductRepository by lazy { ProductRepositoryImpl() }
+}
