+++ app/src/main/java/woowacourse/shopping/ui/MainViewModel.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.ui
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import woowacourse.shopping.data.CartRepository
+import woowacourse.shopping.domain.CartRepository
 import woowacourse.shopping.domain.Product
 import woowacourse.shopping.domain.ProductRepository
